+++ exchangelib/fields.py
@@ -951,3 +951,13 @@ class ItemField(FieldURIField):
     def to_xml(self, value, version):
         # We don't want to wrap in an Item element
         return value.to_xml(version=version)
+
+
+class EffectiveRightsField(EWSElementField):
+    def __init__(self, *args, **kwargs):
+        from .properties import EffectiveRights
+        kwargs['value_cls'] = EffectiveRights
+        super(EffectiveRightsField, self).__init__(*args, **kwargs)
+
+    def from_xml(self, elem, account):
+        return super(EffectiveRightsField, self).from_xml(elem=elem, account=account)